+++ features/scripts/build_android_app.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 set -o errexit
 
+# Use fvm if it is on the PATH
+if [ `command -v fvm` ]; then
+  FLUTTER_BIN="fvm flutter"
+fi
+
 if [ -z "$FLUTTER_BIN" ]; then
   FLUTTER_BIN="flutter"
 fi