+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 4.2.1 (2026-04-29)
+
+- Added `isStarted` to `BugsnagClient` to allow checking whether Bugsnag has been initialized via `start` or `attach`, helping prevent method calls before initialization. [#312](https://github.com/bugsnag/bugsnag-flutter/pull/312)
+
+## TBD
+
+- Upgrade Android compileSdkVersion from 31 to 34 [#305](https://github.com/bugsnag/bugsnag-flutter/pull/305)
+- Amend secondary instance URL to bugsnag.smartbear.com [#299](https://github.com/bugsnag/bugsnag-flutter/pull/299)
+
 ## 4.2.0 (2025-09-29)
 
 - Added `groupingDiscriminator` property to events and methods for setting and getting the global `groupingDiscriminator` to `bugsnag` [#294](https://github.com/bugsnag/bugsnag-flutter/pull/294)